feature(29#,28)

+++ src/main/java/ru/skypro/homework/repository/UserRepository.java
@@ -8,8 +8,10 @@ import ru.skypro.homework.model.User;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findUserByEmailIgnoreCase(String email);
+public interface UserRepository extends JpaRepository<User, Integer> {
+
+        Optional<User> findUserByEmailIgnoreCase(String email);
+
     }
